refactor: 관심사 view 컴포넌트 구조 리팩토링
