+++ src/views/interest/components/select-interest-image/select-interest-image.tsx
