+++ src/views/interest/components/select-interest/select-interest-item.tsx
